test(reorganize): reorganize `tests/clipper/clipper_test.go` for `cli/clipper/clipper.go` tests
Reorganized `tests/clipper/clipper_test.go` to focus exclusively on testing functionalities within `cli/clipper/clipper.go`, ensuring clear and specific test coverage.
